autogen/pyproject.toml
[project]
name = "autogen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "dacite>=1.9.2",
    "jinja2>=3.1.6",
    "libclang>=18.1.1",
    "types-clang>=0.14.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
